requirements.txt
aiofile==3.5.0
caio==0.8.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
motor==2.4.0
pydantic==1.8.1
pymongo==3.11.3
PyYAML==5.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
